Remove treemacs-core.el, put autoloads into treemacs.el.
The modularization scheme did not work - with autoloads placed into core.el only
the part required in core.el was loaded, as opposed the entire project being
required in treemacs.el. Notably all the faces were missing. Going back to the
old way will fix that.
Another solution would've probably been to place autoloads on everything, but
that is 1) leaky, since you always need to look out for gaps, and 2) leaves the
broken loading scheme intact.
As of right now we're going back to treemacs.el being the *only* autoload entry
point from where the rest of the project is guaranteed to load. This makes it
impossible to require treemacs.el from elsewhere (the reason for adding core.el
in the first place), but that's a reasonable price - warnings are easy to
suppress.