Do not recursively pass down git state when reopening dirs.
The recent change to ask for the git status of the current folder only, while
massively improving performance when opening local nodes on a repo with many
changes, means the git state cannot be reused further down the folder chain.
Given the expected small size of localized changes, and their fetching being
async, this should still be a worthwhile trade off.