Make sure symlinks are resolved for git-mode.
Treemacs will now always resolve symlinks in extended git-mode, making it
possible to to see the git status of linked directories that are under version
control in their real position.

This does not apply to files, symlinks are resolved only when expanding
directories.

*Not* resolving symlinks is likewise not possible since git will do it under the
hood anyway (at least it seems to do so by default), so the files linked *into*
a repository will never show up as new in treemacs, or git itself for that
matter. All treemacs can do is correctly interpret the status output by git.

Fixes #210.