[Persistence] Prompt before removing a persisted project
When the project directory does not exist, don't unilaterally decide
that it's no longer needed. By default, ask the user what to do with
completing-read. The configuration option
treemacs-missing-project-action was added to configure this behavior.

Additionally, treemacs--persist is removed from kill-emacs-hook
during restore as not to save an empty persistence file if the user
quits during completing read (or somehow else manages to quit Emacs
during treemacs--restore).