[Git] Improve treemacs-git-status.py performance
String concatenation in Python causes a huge amount of GC and memory
allocations, making Python spend most of its time in the brk syscall.
This makes output a list, and only joins it at the end, avoiding GC of
the previous strings.

Before the change, outputting a hash table consisting of 22204 entries
took 8 seconds and 20418 brk syscalls, after it takes 19 brks and 317
milliseconds.