[Docs] Update examples to adhere to use-package semantics
In the `use-package` documentation it suggested that when using the `:after`
keyword with multiple packages, they should be in a list. This have become
apparent to me after working with `treemacs` for sometime and getting certain
errors. I updated the example in the doc, just in-case other users copy the
examples and try to use it directly without modification, then it will just work
without problems.

If applied, this commit will:

    - Update the example that uses use-package :after key word with multiple
      packages to be a list.