Fix emacs native compilation warnings
To reproduce, have emacs built with native compilation and notice the
compilation logs. You can then open the offending file and run `M-x
emacs-lisp-native-compile-and-load` before and after the changes to see
the warning is removed.

```
Warning: ‘point-at-bol’ is an obsolete function (as of 29.1); use ‘line-beginning-position’ or ‘pos-bol’ instead.
Warning: ‘point-at-eol’ is an obsolete function (as of 29.1); use ‘line-end-position’ or ‘pos-eol’ instead.
```