fix: don't set window parameters if treemacs buff has no window
This fixes bug #1149 on the treemacs issue list.  This occurs when
treemacs-mode is called on a buffer that doesn't have a window.  This
occurs in the lsp hooks when a lsp event occurs, and the lsp buffer in
question (in my case, the error list) doesn't have a window.  The
result is whatever window that was selected gets
'no-delete-other-windows set on it.  Afterwards C-x 1 doesn't work as
expected.

https://github.com/Alexander-Miller/treemacs/issues/1149